break the index.spec into multiple files one per fn
