+++ test/isDef.spec.ts
@@ -1,0 +1,18 @@
+import { isDef } from "../src"
+
+describe('isDef', () => {
+    it('returns true if the provided value is defined', () => {
+        expect(isDef('Hej')).toBe(true)
+        expect(isDef(13)).toBe(true)
+        expect(isDef(false)).toBe(true)
+        expect(isDef(NaN)).toBe(true)
+        expect(isDef({})).toBe(true)
+        expect(isDef([])).toBe(true)
+    })
+
+    it('returns false if the provided value is not defined', () => {
+        // @ts-expect-error
+        expect(isDef()).toBe(false)
+        expect(isDef(undefined)).toBe(false)
+    })
+})
