+++ test/isObj.spec.ts
@@ -1,0 +1,15 @@
+import { isObj } from '../src'
+
+describe('isObj', () => {
+    it('returns true for an object', () => {
+        expect(isObj({})).toBe(true)
+    })
+
+    it('returns true for an array', () => {
+        expect(isObj([])).toBe(true)
+    })
+
+    it('returns false for null', () => {
+        expect(isObj(null)).toBe(false)
+    })
+})
